+++ akka-http-core/src/test/scala/akka/http/scaladsl/settings/SettingsEqualitySpec.scala
@@ -11,18 +11,7 @@ import org.scalatest.WordSpec
 
 class SettingsEqualitySpec extends WordSpec with Matchers {
 
-  val config = ConfigFactory.parseString("""
-    akka.http.routing {
-      verbose-error-messages = off
-      file-get-conditional = on
-      render-vanity-footer = yes
-      range-coalescing-threshold = 80
-      range-count-limit = 16
-      decode-max-bytes-per-chunk = 1m
-      decode-max-size = 8m
-      file-io-dispatcher = ${akka.stream.blocking-io-dispatcher}
-    }
-  """).withFallback(ConfigFactory.load).resolve
+  val config = ConfigFactory.load.resolve
 
   "equality" should {
     "hold for ConnectionPoolSettings" in {
@@ -47,14 +36,6 @@ class SettingsEqualitySpec extends WordSpec with Matchers {
 
       s1 shouldBe s2
       s1.toString should startWith("ClientConnectionSettings(")
-    }
-
-    "hold for RoutingSettings" in {
-      val s1 = RoutingSettings(config)
-      val s2 = RoutingSettings(config)
-
-      s1 shouldBe s2
-      s1.toString should startWith("RoutingSettings(")
     }
 
     "hold for ServerSettings" in {
